+++ option-impl.ts
@@ -1,5 +1,10 @@
-import { IOption, None, Option as OptionType, Some } from "./option.ts";
-import { Err, Ok, Result } from "./result.ts";
+import {
+    type IOption,
+    None,
+    type Option as OptionType,
+    Some,
+} from "./option.ts";
+import { Err, Ok, type Result } from "./result.ts";
 
 class Option<A> implements IOption<A> {
     readonly result!: boolean;